Revert broken changes
+ Add Eclipse project files
+ Add JavaFX package for graphical interface

+++ src/main/Static.java
@@ -72,7 +72,7 @@ public final class Static {
 	
 	public static final void sleep(int milliseconds) {
 		try {
-			Thread.sleep((long) milliseconds);
+			Thread.sleep(milliseconds);
 		} catch (InterruptedException e) {
 			Main.onerror(e);
 		}
